[mod]修复品牌控制器修改保存上传图片

+++ app/Http/Controllers/Admin/BrandController.php
@@ -116,16 +116,17 @@ class BrandController extends Controller
         //接收数据 加入model
         $model->setRawAttributes($request->only(['name','desc']));
 
+        if ($id != -1 && $request->del) {
+            $model->img = null;
+            $model->thumb = null;
+            IQuery::destroyPic(new Brand, $id, 'img');
+        }
+
         //资源、上传图片名称、是否生成缩略图
         $imgs = IQuery::upload($request,'img',true);
         if ($imgs != 'false') {
             $model->img = $imgs['pic'];
             $model->thumb = $imgs['thumb'];
-        }
-        if ($id != -1) {
-            if ($imgs != 'false' || $request->del) {
-                IQuery::destroyPic(new Brand, $id, 'img');
-            }
         }
 
         if ($model->save()) {
